src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/get-destination-insertion-points-skips-user-agent-shadow-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests that getDestinationInsertionPoints() should skip insertion points in user-agent shadow roots.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
